Changed to save when starting a presentation
