+++ components/edit/AutoSavePage.tsx
@@ -6,19 +6,14 @@
  *
  * Copyright (C) 2021 hello-slide
  **********************************************************/
-import {useToast} from '@chakra-ui/react';
 import {useRouter} from 'next/router';
 import React from 'react';
-import {useRecoilValue, useRecoilState} from 'recoil';
-import SetPage from '../../utils/api/setPage';
+import useSavePage from '../../hooks/useSavePage';
 import {removeBeforeUnLoad} from '../../utils/event/beforeUnLoad';
-import {PageDataState, NowPageDataState} from '../../utils/state/atom';
 
 const AutoSave = () => {
-  const [pageData, setPageData] = useRecoilState(PageDataState);
-  const nowPageData = useRecoilValue(NowPageDataState);
-  const toast = useToast();
   const router = useRouter();
+  const [, setPage] = useSavePage();
 
   const [nowPath, setNowPath] = React.useState<string>(undefined);
 
@@ -32,23 +27,6 @@ const AutoSave = () => {
       setNowPath(router.asPath);
     }
   }, [router.asPath]);
-
-  const setPage = () => {
-    if (typeof pageData !== 'undefined') {
-      const pageId = pageData.id;
-      const setPageAPI = new SetPage();
-
-      setPageAPI.run(nowPageData?.id, pageId, pageData).catch(error => {
-        toast({
-          title: 'ページを保存できませんでした。',
-          description: `${error}`,
-          status: 'error',
-        });
-      });
-
-      setPageData(undefined);
-    }
-  };
 
   return <></>;
 };
